agent-service/requirements.txt
# Horiz Quant Researcher — Agent Service Dependencies
claude-agent-sdk>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.0
python-dotenv>=1.0.0
